tools/social_media_tool.py
import requests
import json
import schedule
import time
from datetime import datetime, timedelta
from typing import Dict, Any, List
from crewai_tools import BaseTool
from config import Config

class PostToSocialTool(BaseTool):
    name: str = "Post to Social Media Tool"
    description: str = "Post or schedule videos to TikTok, Instagram Reels, and YouTube Shorts with platform-specific optimization."
    
    def _run(self, video_path: str, caption: str, platforms: List[str] = None, schedule_time: str = None) -> str:
        """
        Post or schedule video to social media platforms.
        
        Args:
            video_path (str): Path to the video file
            caption (str): Caption for the post
            platforms (List[str]): List of platforms to post to
            schedule_time (str): Optional schedule time (ISO format)
            
        Returns:
            str: Posting results for each platform
        """
        try:
            if platforms is None:
                platforms = Config.PLATFORMS
            
            results = []
            
            for platform in platforms:
                if schedule_time:
                    result = self._schedule_post(video_path, caption, platform, schedule_time)
                else:
                    result = self._post_immediately(video_path, caption, platform)
                
                results.append(f"{platform.upper()}: {result}")
            
            return "\n".join(results)
            
        except Exception as e:
            return f"Error posting to social media: {str(e)}"
    
    def _post_immediately(self, video_path: str, caption: str, platform: str) -> str:
        """Post video immediately to specified platform."""
        # For demo purposes, return mock results
        if "captioned_video_" in video_path or "example.com" in video_path:
            return self._mock_post(video_path, caption, platform)
        
        if platform == "tiktok":
            return self._post_to_tiktok(video_path, caption)
        elif platform == "instagram":
            return self._post_to_instagram(video_path, caption)
        elif platform == "youtube_shorts":
            return self._post_to_youtube_shorts(video_path, caption)
        else:
            return f"Unsupported platform: {platform}"
    
    def _post_to_tiktok(self, video_path: str, caption: str) -> str:
        """Post video to TikTok."""
        try:
            # TikTok API endpoint (simplified)
            url = "https://open-api.tiktok.com/share/video/upload/"
            
            headers = {
                "Authorization": f"Bearer {Config.TIKTOK_ACCESS_TOKEN}",
                "Content-Type": "application/json"
            }
            
            # Prepare video upload
            with open(video_path, 'rb') as video_file:
                files = {'video': video_file}
                data = {
                    'text': f"{caption} {Config.BRAND_HANDLE}",
                    'privacy_level': 'SELF_ONLY',  # Start with private for testing
                    'disable_duet': False,
                    'disable_comment': False,
                    'disable_stitch': False,
                    'brand_content_toggle': False
                }
                
                response = requests.post(url, headers=headers, data=data, files=files)
                
                if response.status_code == 200:
                    result = response.json()
                    return f"Posted successfully! Video ID: {result.get('video_id', 'N/A')}"
                else:
                    return f"Error posting: {response.status_code}"
                    
        except Exception as e:
            return f"TikTok posting error: {str(e)}"
    
    def _post_to_instagram(self, video_path: str, caption: str) -> str:
        """Post video to Instagram Reels."""
        try:
            # Instagram Basic Display API
            url = f"https://graph.instagram.com/v17.0/me/media"
            
            params = {
                'media_type': 'VIDEO',
                'video_url': video_path,  # Assuming video is already uploaded to accessible URL
                'caption': f"{caption} {Config.BRAND_HANDLE}",
                'access_token': Config.INSTAGRAM_ACCESS_TOKEN
            }
            
            response = requests.post(url, params=params)
            
            if response.status_code == 200:
                result = response.json()
                media_id = result.get('id')
                
                # Publish the media
                publish_url = f"https://graph.instagram.com/v17.0/me/media_publish"
                publish_params = {
                    'creation_id': media_id,
                    'access_token': Config.INSTAGRAM_ACCESS_TOKEN
                }
                
                publish_response = requests.post(publish_url, params=publish_params)
                
                if publish_response.status_code == 200:
                    return f"Posted successfully! Media ID: {media_id}"
                else:
                    return f"Error publishing: {publish_response.status_code}"
            else:
                return f"Error creating media: {response.status_code}"
                
        except Exception as e:
            return f"Instagram posting error: {str(e)}"
    
    def _post_to_youtube_shorts(self, video_path: str, caption: str) -> str:
        """Post video to YouTube Shorts."""
        try:
            # YouTube Data API v3
            url = "https://www.googleapis.com/upload/youtube/v3/videos"
            
            headers = {
                "Authorization": f"Bearer {Config.YOUTUBE_API_KEY}",
                "Content-Type": "application/json"
            }
            
            metadata = {
                "snippet": {
                    "title": f"{caption[:100]}...",  # Truncate for title
                    "description": f"{caption}\n\n{Config.BRAND_HANDLE}\n\n#Shorts #Tax #Finance #Baby #Viral",
                    "tags": ["tax", "finance", "baby", "comedy", "shorts", "viral", "mclantax"],
                    "categoryId": "23"  # Comedy category
                },
                "status": {
                    "privacyStatus": "private"  # Start with private
                }
            }
            
            with open(video_path, 'rb') as video_file:
                files = {'video': video_file}
                data = {'snippet': json.dumps(metadata)}
                
                response = requests.post(url, headers=headers, data=data, files=files)
                
                if response.status_code == 200:
                    result = response.json()
                    return f"Posted successfully! Video ID: {result.get('id', 'N/A')}"
                else:
                    return f"Error posting: {response.status_code}"
                    
        except Exception as e:
            return f"YouTube posting error: {str(e)}"
    
    def _schedule_post(self, video_path: str, caption: str, platform: str, schedule_time: str) -> str:
        """Schedule a post for later."""
        try:
            # Parse schedule time
            scheduled_datetime = datetime.fromisoformat(schedule_time)
            
            # For demo purposes, just return scheduled confirmation
            return f"Scheduled for {scheduled_datetime.strftime('%Y-%m-%d %H:%M:%S')}"
            
        except Exception as e:
            return f"Error scheduling: {str(e)}"
    
    def _mock_post(self, video_path: str, caption: str, platform: str) -> str:
        """Mock posting for testing purposes."""
        post_id = f"{platform}_{int(time.time())}"
        
        platform_features = {
            "tiktok": {
                "hashtags": "#fyp #viral #tax #baby #comedy #foryou",
                "engagement": "12.5K views, 890 likes, 45 comments",
                "url": f"https://tiktok.com/@mclantax/video/{post_id}"
            },
            "instagram": {
                "hashtags": "#reels #viral #tax #baby #comedy #explore",
                "engagement": "8.2K views, 654 likes, 32 comments",
                "url": f"https://instagram.com/p/{post_id}"
            },
            "youtube_shorts": {
                "hashtags": "#shorts #tax #finance #baby #comedy",
                "engagement": "15.3K views, 1.2K likes, 67 comments",
                "url": f"https://youtube.com/shorts/{post_id}"
            }
        }
        
        features = platform_features.get(platform, platform_features["tiktok"])
        
        return f"""
Posted successfully to {platform.upper()}!

Video: {video_path}
Caption: {caption[:50]}...
Post ID: {post_id}
URL: {features['url']}
Hashtags: {features['hashtags']}
Brand Tag: {Config.BRAND_HANDLE}

Expected Engagement: {features['engagement']}
Status: Live and ready for viral growth!
""" 