requirements.txt
crewai>=0.28.8
crewai[tools]>=0.28.8
requests>=2.31.0
python-dotenv>=1.0.0
pydantic>=2.5.0
langchain>=0.1.0
langchain-openai>=0.0.5
openai>=1.12.0
ffmpeg-python>=0.2.0
pillow>=10.0.0
moviepy>=1.0.3
flask>=2.3.0
flask-cors>=4.0.0 